refactor(feature-user): update `user-edit` feature module

+++ feature/mobile/user-edit/src/main/kotlin/com/flixclusive/feature/mobile/user/edit/UserEditViewModel.kt
@@ -1,21 +1,23 @@
 package com.flixclusive.feature.mobile.user.edit
 
 import androidx.lifecycle.ViewModel
+import com.flixclusive.core.common.dispatchers.AppDispatchers
+import com.flixclusive.core.common.locale.UiText
 import com.flixclusive.core.database.entity.user.User
 import com.flixclusive.core.datastore.DataStoreManager
-import com.flixclusive.core.strings.UiText
-import com.flixclusive.core.util.coroutines.AppDispatchers.Companion.launchOnIO
-import com.flixclusive.data.library.recent.WatchHistoryRepository
-import com.flixclusive.data.library.watchlist.WatchlistRepository
-import com.flixclusive.domain.database.repository.SearchHistoryRepository
-import com.flixclusive.domain.database.repository.UserRepository
-import com.flixclusive.domain.provider.ProviderUnloaderUseCase
-import com.flixclusive.domain.provider.repository.ProviderRepository
-import com.flixclusive.domain.session.UserSessionManager
+import com.flixclusive.data.database.repository.SearchHistoryRepository
+import com.flixclusive.data.database.repository.UserRepository
+import com.flixclusive.data.database.repository.WatchProgressRepository
+import com.flixclusive.data.database.repository.WatchlistRepository
+import com.flixclusive.data.database.session.UserSessionManager
+import com.flixclusive.data.provider.repository.ProviderRepository
+import com.flixclusive.domain.provider.usecase.manage.UnloadProviderUseCase
 import com.flixclusive.feature.mobile.user.edit.OnRemoveNavigationState.Companion.getStateIfUserIsLoggedIn
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 import com.flixclusive.core.strings.R as LocaleR
 
@@ -28,15 +30,23 @@ internal class UserEditViewModel
         private val userSessionManager: UserSessionManager,
         private val searchHistoryRepository: SearchHistoryRepository,
         private val watchlistRepository: WatchlistRepository,
-        private val watchHistoryRepository: WatchHistoryRepository,
+        private val watchProgressRepository: WatchProgressRepository,
         private val providerRepository: ProviderRepository,
-        private val providerUnloaderUseCase: ProviderUnloaderUseCase,
+        private val unloadProvider: UnloadProviderUseCase,
+        private val appDispatchers: AppDispatchers,
     ) : ViewModel() {
         private val _onRemoveNavigationState = MutableSharedFlow<OnRemoveNavigationState>()
         val onRemoveNavigationState = _onRemoveNavigationState.asSharedFlow()
 
+        private var removeUserJob: Job? = null
+        private var editUserJob: Job? = null
+        private var clearSearchHistoryJob: Job? = null
+        private var clearLibrariesJob: Job? = null
+
         fun onRemoveUser(userId: Int) {
-            launchOnIO {
+            if (removeUserJob?.isActive == true) return
+
+            removeUserJob = appDispatchers.ioScope.launch {
                 val isUserLoggedIn = isUserLoggedIn(userId)
                 val navigationState = getStateIfUserIsLoggedIn(isUserLoggedIn)
 
@@ -48,7 +58,7 @@ internal class UserEditViewModel
 
                 // CLEAR DATABASE
                 searchHistoryRepository.clearAll(userId)
-                watchHistoryRepository.removeAll(userId)
+                watchProgressRepository.removeAll(userId)
                 watchlistRepository.removeAll(userId)
                 userRepository.deleteUser(userId)
                 // ===
@@ -58,13 +68,17 @@ internal class UserEditViewModel
         }
 
         fun onEditUser(user: User) {
-            launchOnIO {
+            if (editUserJob?.isActive == true) return
+
+            editUserJob = appDispatchers.ioScope.launch {
                 userRepository.updateUser(user)
             }
         }
 
         fun onClearSearchHistory(userId: Int) {
-            launchOnIO {
+            if (clearSearchHistoryJob?.isActive == true) return
+
+            clearSearchHistoryJob = appDispatchers.ioScope.launch {
                 searchHistoryRepository.clearAll(ownerId = userId)
             }
         }
@@ -73,11 +87,13 @@ internal class UserEditViewModel
             userId: Int,
             libraries: List<Library>,
         ) {
-            launchOnIO {
+            if (clearLibrariesJob?.isActive == true) return
+
+            clearLibrariesJob = appDispatchers.ioScope.launch {
                 libraries.forEach {
                     when (it) {
                         is Library.Watchlist -> watchlistRepository.removeAll(userId)
-                        is Library.WatchHistory -> watchHistoryRepository.removeAll(userId)
+                        is Library.WatchHistory -> watchProgressRepository.removeAll(userId)
                         is Library.CustomList -> throw IllegalStateException("Custom libraries are not yet implemented")
                     }
                 }
@@ -87,7 +103,7 @@ internal class UserEditViewModel
         private suspend fun clearProviders(userId: Int) {
             if (!isUserLoggedIn(userId)) return
             providerRepository.getProviders().forEach {
-                providerUnloaderUseCase.unload(it)
+                unloadProvider(it)
             }
         }
